Add websocket subscriber map
This map will keep track of which sink clients are
subscribed to which source clients. I will have to define
a protocol by which multiple types of messages can be sent
- I know I will at least need commands for clients to
register and unregister as a source, and for clients to
subscribe and unsubcribe from sources, and of course for a
source to send the ECG data.

Also I will probably want each source to provide a display
name when registering, and receive a UUID. Then when the
page is loaded there will be an endpoint to get all sources,
and they will be represented to the user with the display
name but internally represented by the UUID.

I think the correct way to implement this is to make a data
class for each type of command, and then use a library like
kotlinx.serialization or klaxon to turn the strings into
objects. The need to rigidly define these message formats
feels like a downside of using a language like Kotlin,
as opposed to Python/Clojure/JS which let you freely
turn a JSON string into a "map of anything->anything".