Add data classes for parsing JSON commands
The kotlinx.serialization library lets us automatically
parse serializable classes to and from JSON, which should
be helpful for sending and receiving commands. I expect
the most annoying part will be finding out what kind of
command we have - I think the only solution will be to
try to parse it as each type of command, and then handle
whichever kind it turns out to be.