Add function to handle websocket subscriptions
I'm starting to worry about the speed of the JSON parsing. It
seems to be very slow to "fail" when the string is not of the
correct format, and the current design would have that happen
often, since it just tries each format in turn when parsing
a message whose type it does not know. If I don't find another
solution I may have to switch off of JSON to some other string
format. It would be fairly easy to replace "{DATA:FOO}" with
"DATA:FOO" and parse it with basic string functions, as long
as I know my data won't be that complicated.