Add function to handle websocket unsubscriptions
I found something interesting about the slow JSON decoding,
which is that it seems like the failure is only slow the
first time it happens. Maybe there is some kind of overhead
within the library that is only apparent on the first
call? I will look into a way to get the overhead out of
the way when the server starts, instead of having to put
that on users.