Make clients auto-unsubscribe when closing
Also apparently clients don't throw the exception I
expected them to when closing. For now I catch Exception
to detect when they close, but I would like to fix it
so that the specific exception is caught. Maybe the old
exception type is thrown for a manual close, and there
is something else thrown when the client websocket
"crashes", ex. when the page is closed.