Rework data structure for registered websockets
Now, websockets are automatically "registered" when they connect.
This prevents us from having some websockets that exist but are
not registered. To set their name, clients use the new name
command, which works similarly to the old register command.

The new layout for the data makes it a lot easier to push data
to all clients, which happens fairly often since each new
client or client renaming makes a publish happen.